Move interfaces to where they are used.
